small tweaks to reorganize ethers utiltieis
