+++ libs/utilities/common-utilities/src/utils/validation.ts
@@ -1,20 +1,4 @@
-/*
-  START react-hook-form REMOVE DEPENDENCY
-  taken from Validate type in react-hook-form
-
-  removing import to remove dependency on react-hook-form:
-  import { Validate } from 'react-hook-form';
-*/
-export declare type Message = string;
-export declare type ValidateResult = Message | Message[] | boolean | undefined;
-export declare type Validate<TFieldValue> = (
-  value: TFieldValue
-) => ValidateResult | Promise<ValidateResult>;
-
-/*
-  END react-hook-form REMOVE DEPENDENCY
-*/
-
+import { Validate } from './remove-dependencies';
 import { isArray, isBoolean, isNumberish, isString } from './typeguards';
 
 export const ValErrMsgs = {
