+++ libs/utilities/common-utilities/tsconfig.json
@@ -9,7 +9,8 @@
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
     "allowJs": true,
-    "declaration": true
+    "declaration": true,
+    "composite": true
   },
   "files": [],
   "include": [],
